* src/jtag.h: Lengthen JTAG_RESPONSE_TIMEOUT so slower systems have a fair chance to respond. * src/remote.cc: Passing an int to read(2) is not bigendian friendly, so pass an unsigned char instead.
git-svn-id: svn://svn.code.sf.net/p/avarice/code/trunk@14 a7e36acc-7495-4a18-91ca-7b9a7881be12

+++ avarice/src/remote.cc
@@ -116,7 +116,7 @@ static void waitForGdbInput(void)
     exit cleanly if EOF detected on gdbFileDescriptor. **/
 int getDebugChar(void)
 {
-    int c = (int)0;
+    uchar c = 0;
     int result;
 
     do
@@ -140,7 +140,7 @@ int getDebugChar(void)
 
 int checkForDebugChar(void)
 {
-    int c = (int)0;
+    uchar c = 0;
     int result;
 
     result = read(gdbFileDescriptor, &c, 1);
